requirements.txt
anyio==4.12.0
certifi==2025.11.12
charset-normalizer==3.4.4
exceptiongroup==1.3.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
notion-client==2.7.0
numpy==2.0.2
pandas==2.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
six==1.17.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.0
